+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 * New `read_scanimage_metadata` for reading scanimage metadata from a file directly as a python dict [#405](https://github.com/catalystneuro/roiextractors/pull/401)
+* Added `ScanImageImagingExtractor` for simplifying reading ScanImage data [#412](https://github.com/catalystneuro/roiextractors/pull/412)
 
 ### Fixes
 * Use `SI.hChannels.channelSave` or `SI.hChannels.channelsave` to determine number of channels for ScanImage extractors when available [#401](https://github.com/catalystneuro/roiextractors/pull/401)